% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_functions.R
\name{calculate_ic_from_extracted}
\alias{calculate_ic_from_extracted}
\title{Internal function to calculate IC from extracted clustering array}
\usage{
calculate_ic_from_extracted(clustering_array)
}
\description{
Internal function to calculate IC from extracted clustering array
}
\keyword{internal}
