% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_core.R
\name{filter_problematic_clusters}
\alias{filter_problematic_clusters}
\title{Filter out cluster numbers that produce inconsistent results}
\usage{
filter_problematic_clusters(
  gamma_dict,
  cluster_range,
  igraph_obj,
  objective_function,
  remove_threshold,
  n_workers,
  verbose
)
}
\description{
Filter out cluster numbers that produce inconsistent results
}
\keyword{internal}
