% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{get_sigev}
\alias{get_sigev}
\title{Helper function: Get significant eigenvalues using RMT
Julia equivalent: get_sigev(X,Xr;device="gpu")}
\usage{
get_sigev(X, Xr)
}
\arguments{
\item{X}{Scaled data matrix}

\item{Xr}{Randomized data matrix}
}
\value{
List with signal eigenvalues, eigenvectors, and RMT parameters
}
\description{
Splits eigencomponents into signal and noise using RMT criteria.
}
\details{
This routine combines MP filtering and Tracy-Widom thresholding to identify
signal eigenvalues/eigenvectors from observed data relative to a randomized
null matrix, and returns both signal and noise components for downstream
robustness analysis.
}
\keyword{internal}
