% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_wrapper.R
\name{is.weighted.igraph}
\alias{is.weighted.igraph}
\title{Check if igraph object is weighted}
\usage{
is.weighted.igraph(igraph_obj)
}
\arguments{
\item{igraph_obj}{igraph object}
}
\value{
Logical indicating if graph is weighted
}
\description{
Check if igraph object is weighted
}
\keyword{internal}
