% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_wrapper.R
\name{leiden_clustering_alternative}
\alias{leiden_clustering_alternative}
\title{Alternative leiden clustering using leiden package if available}
\usage{
leiden_clustering_alternative(
  adjacency_matrix,
  resolution,
  objective_function,
  n_iterations,
  beta,
  initial_membership = NULL
)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}

\item{resolution}{Resolution parameter}

\item{objective_function}{Objective function}

\item{n_iterations}{Number of iterations}

\item{beta}{Beta parameter}

\item{initial_membership}{Initial membership}
}
\value{
Cluster assignments
}
\description{
Alternative leiden clustering using leiden package if available
}
\keyword{internal}
