% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_core.R
\name{load_cluster_matrix}
\alias{load_cluster_matrix}
\title{Load clustering matrix from reference}
\usage{
load_cluster_matrix(matrix_ref)
}
\arguments{
\item{matrix_ref}{Matrix reference list}
}
\value{
Matrix
}
\description{
Materializes a matrix previously returned by \code{store_cluster_matrix()}.
}
\keyword{internal}
