% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{mp_calculation}
\alias{mp_calculation}
\title{Helper function: Marchenko-Pastur calculation
Julia equivalent: _mp_calculation(L, Lr, eta=1, eps=1e-6, max_iter=10000)}
\usage{
mp_calculation(L, Lr, eta = 1, eps = 1e-06, max_iter = 10000)
}
\arguments{
\item{L}{Original eigenvalues}

\item{Lr}{Random eigenvalues}

\item{eta}{Learning rate}

\item{eps}{Convergence threshold}

\item{max_iter}{Maximum iterations}
}
\value{
List with filtered eigenvalues and bounds
}
\description{
Helper function: Marchenko-Pastur calculation
Julia equivalent: _mp_calculation(L, Lr, eta=1, eps=1e-6, max_iter=10000)
}
