% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{proj_l}
\alias{proj_l}
\title{Helper function: Project to simplex (L1 normalization)
Julia equivalent: proj_l = x -> issparse(x) ? spdiagm(1 ./sum(x,dims=2)[:]) * x : x ./ sum(x,dims=2)}
\usage{
proj_l(x)
}
\arguments{
\item{x}{Matrix to normalize}
}
\value{
L1 normalized matrix
}
\description{
Helper function: Project to simplex (L1 normalization)
Julia equivalent: proj_l = x -> issparse(x) ? spdiagm(1 ./sum(x,dims=2)[:]) * x : x ./ sum(x,dims=2)
}
