% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_runtime.R
\name{release_cluster_matrix_refs}
\alias{release_cluster_matrix_refs}
\title{Release multiple matrix references}
\usage{
release_cluster_matrix_refs(matrix_refs)
}
\arguments{
\item{matrix_refs}{List of matrix references}
}
\value{
Invisibly TRUE
}
\description{
Convenience helper to release a list of references returned by
\code{store_cluster_matrix()}.
}
\keyword{internal}
