% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_core.R
\name{clustering_main}
\alias{clustering_main}
\title{Core clustering algorithm implementing binary search and optimization}
\usage{
clustering_main(
  igraph_obj,
  cluster_range,
  n_workers = max(1, parallel::detectCores() - 1),
  n_trials,
  n_bootstrap,
  seed = NULL,
  beta,
  n_iterations,
  max_iterations,
  objective_function,
  remove_threshold,
  resolution_tolerance,
  verbose,
  in_parallel_context = FALSE
)
}
\arguments{
\item{igraph_obj}{igraph object to cluster}

\item{cluster_range}{Vector of cluster numbers to test}

\item{n_workers}{Number of parallel workers (default: max(1, parallel::detectCores() - 1))}

\item{n_trials}{Number of clustering trials per resolution}

\item{n_bootstrap}{Number of bootstrap iterations}

\item{beta}{Beta parameter for Leiden clustering}

\item{n_iterations}{Number of Leiden iterations}

\item{max_iterations}{Maximum iterations for optimization}

\item{objective_function}{Objective function ("modularity" or "CPM")}

\item{remove_threshold}{Threshold for removing inconsistent results}

\item{resolution_tolerance}{Tolerance for resolution parameter search}

\item{verbose}{Whether to print progress messages}

\item{in_parallel_context}{Whether this function is called from within a parallel context (default: FALSE)}
}
\value{
List with clustering results
}
\description{
Core clustering algorithm implementing binary search and optimization
}
\keyword{internal}
