% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_functions.R
\name{extract_clustering_array}
\alias{extract_clustering_array}
\title{Extract unique clustering arrays and their probabilities}
\usage{
extract_clustering_array(clustering_matrix)
}
\arguments{
\item{clustering_matrix}{Matrix where each column is a clustering result}
}
\value{
List with unique clustering arrays and their probabilities
}
\description{
Extract unique clustering arrays and their probabilities
}
\keyword{internal}
