% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_core.R
\name{find_resolution_ranges}
\alias{find_resolution_ranges}
\title{Find resolution parameter ranges for each cluster number using binary search}
\usage{
find_resolution_ranges(
  igraph_obj,
  cluster_range,
  start_g,
  end_g,
  objective_function,
  resolution_tolerance,
  n_workers,
  verbose
)
}
\description{
Find resolution parameter ranges for each cluster number using binary search
}
\keyword{internal}
