% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_functions.R
\name{get_best_clustering}
\alias{get_best_clustering}
\title{Get the best clustering from extracted clustering array}
\usage{
get_best_clustering(clustering_array)
}
\arguments{
\item{clustering_array}{Result from extract_clustering_array}
}
\value{
Best clustering labels
}
\description{
Get the best clustering from extracted clustering array
}
\keyword{internal}
