% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{get_eigen}
\alias{get_eigen}
\title{Helper function: Compute eigendecomposition
Julia equivalent: _get_eigen(Y;device="cpu")}
\usage{
get_eigen(Y)
}
\arguments{
\item{Y}{Symmetric matrix}
}
\value{
List with eigenvalues and eigenvectors
}
\description{
Helper function: Compute eigendecomposition
Julia equivalent: _get_eigen(Y;device="cpu")
}
