% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{get_eigvec}
\alias{get_eigvec}
\title{Helper function: Get eigenvalues and eigenvectors
Julia equivalent: get_eigvec(X;device="gpu")}
\usage{
get_eigvec(X)
}
\arguments{
\item{X}{Input matrix}
}
\value{
List with eigenvalues and eigenvectors
}
\description{
Helper function: Get eigenvalues and eigenvectors
Julia equivalent: get_eigvec(X;device="gpu")
}
