% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_wrapper.R
\name{graph_to_igraph}
\alias{graph_to_igraph}
\title{Convert Seurat graph to igraph object}
\usage{
graph_to_igraph(seurat_graph)
}
\arguments{
\item{seurat_graph}{Graph from Seurat object (typically SNN or KNN graph)}
}
\value{
igraph object
}
\description{
Convert Seurat graph to igraph object
}
\keyword{internal}
