% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_wrapper.R
\name{leiden_clustering}
\alias{leiden_clustering}
\title{Perform Leiden clustering on an igraph object}
\usage{
leiden_clustering(
  igraph_obj,
  resolution,
  objective_function,
  n_iterations,
  beta,
  initial_membership = NULL
)
}
\arguments{
\item{igraph_obj}{igraph object to cluster}

\item{resolution}{Resolution parameter for clustering}

\item{objective_function}{Objective function ("modularity" or "CPM")}

\item{n_iterations}{Number of iterations}

\item{beta}{Beta parameter}

\item{initial_membership}{Initial cluster membership (optional)}
}
\value{
Vector of cluster assignments (0-based)
}
\description{
Perform Leiden clustering on an igraph object
}
\keyword{internal}
