% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{mp_parameters}
\alias{mp_parameters}
\title{Helper function: Marchenko-Pastur parameters
Julia equivalent: _mp_parameters(L)}
\usage{
mp_parameters(L)
}
\arguments{
\item{L}{Vector of eigenvalues}
}
\value{
List of Marchenko-Pastur parameters
}
\description{
Helper function: Marchenko-Pastur parameters
Julia equivalent: _mp_parameters(L)
}
