% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_core.R
\name{optimize_clustering}
\alias{optimize_clustering}
\title{Optimize clustering within a resolution range}
\usage{
optimize_clustering(
  igraph_obj,
  target_clusters,
  gamma_range,
  objective_function,
  n_trials,
  n_bootstrap,
  beta,
  n_iterations,
  max_iterations,
  resolution_tolerance,
  n_workers
)
}
\description{
Optimize clustering within a resolution range
}
\keyword{internal}
