% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{random_nz}
\alias{random_nz}
\title{Helper function: Generate random matrix for null model
Julia equivalent: random_nz(pre_df;rmix=true,mix_p=nothing)}
\usage{
random_nz(X, rmix = TRUE)
}
\arguments{
\item{X}{Sparse matrix}

\item{rmix}{Whether to randomize structure}
}
\value{
Randomized sparse matrix
}
\description{
Helper function: Generate random matrix for null model
Julia equivalent: random_nz(pre_df;rmix=true,mix_p=nothing)
}
