% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{scaled_gdata}
\alias{scaled_gdata}
\title{Helper function: Scale data with centering
Julia equivalent: scaled_gdata(X;dim=1, position_="mean")}
\usage{
scaled_gdata(X, position_ = "mean")
}
\arguments{
\item{X}{Matrix to scale}

\item{position_}{Centering method ("mean" or "median")}
}
\value{
Scaled matrix
}
\description{
Helper function: Scale data with centering
Julia equivalent: scaled_gdata(X;dim=1, position_="mean")
}
