% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{tw_calculation}
\alias{tw_calculation}
\title{Helper function: Tracy-Widom calculation
Julia equivalent: _tw(L, L_mp)}
\usage{
tw_calculation(L, L_mp)
}
\arguments{
\item{L}{Original eigenvalues}

\item{L_mp}{Marchenko-Pastur eigenvalues}
}
\value{
List with Tracy-Widom parameters
}
\description{
Helper function: Tracy-Widom calculation
Julia equivalent: _tw(L, L_mp)
}
