% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{wishart_matrix}
\alias{wishart_matrix}
\title{Helper function: Compute Wishart matrix (covariance matrix)
Julia equivalent: _wishart_matrix(X;device="cpu")}
\usage{
wishart_matrix(X)
}
\arguments{
\item{X}{Input matrix}
}
\value{
Wishart matrix (covariance matrix)
}
\description{
Helper function: Compute Wishart matrix (covariance matrix)
Julia equivalent: _wishart_matrix(X;device="cpu")
}
