% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLENS.R
\name{zscore_with_l2}
\alias{zscore_with_l2}
\title{Helper function: Z-score with L2 normalization
Julia equivalent: zscore_with_l2(X)}
\usage{
zscore_with_l2(X)
}
\arguments{
\item{X}{Matrix to normalize}
}
\value{
Z-score normalized matrix with L2 correction
}
\description{
Helper function: Z-score with L2 normalization
Julia equivalent: zscore_with_l2(X)
}
